+++ src/components/App.test.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./App";
+
 import config from "../config";
+import { App } from "./App";
 
 it("renders without crashing", () => {
   const div = document.createElement("div");