+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- :construction_worker_man: Add application state (redux).
+
 ## 1.0.3 (2019-03-21)
 - :sparkles: Add theme support.
 - :sparkles: Add favicons.